fix: rename doc_ to document_ for consistency
