+++ src/langchain_google_memorystore_redis/__init__.py
@@ -12,8 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from .chat_message_history import MemorystoreChatMessageHistory
-from .doc_loader import MemorystoreDocumentLoader
-from .doc_saver import MemorystoreDocumentSaver
+from .document_loader import MemorystoreDocumentLoader
+from .document_saver import MemorystoreDocumentSaver
 from .vector_store import (
     DistanceStrategy,
     FLATConfig,
